requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
asgiref>=3.7.0

